refactor: extract componet
